requirements.txt
fastapi==0.115.2
uvicorn[standard]
python-multipart==0.0.6
pydantic==2.9.0
pydantic-settings==2.0.0
python-dotenv==1.0.0
PyMuPDF==1.23.1
sentence-transformers==2.3.2
transformers==4.34.0
torch==2.2.3
faiss-cpu==1.7.4
Pillow==10.0.0
openai==1.0.0
huggingface_hub==0.17.1
pytest==8.4.0
httpx==0.26.2
